Add files via upload
1. Write a concise single sentence about yourself as a designer, in under 120 characters in length

Student of Graphic designer with lots of ideas and incorrect humor. Every thing is an inspiration to me.


2. Write a separate short paragraph about who you are, beyond being a designer with a sense of humor; a personal bio, with an emphasis on keywords and under 300 characters in length.

I was born in Czech small city in the mountains, Jilemnice. In my free time I play ice hockey . I like walking around the mountains. I like to spend time with friends at the cottage .  As graphic designer, I like to entertain people with graphic photomontages. 

3. Choose a project to develop as a case study for your portfolio and write about the process using the PAR and RAD content methods. Aim for around 500 words (not characters).
In 2018, I had a phone call from Tomáš Bárta, who is the head of KRPA FORM, a company which (nebo that? ted nevim) focuses on paper production. He told me that they had managed to create a new pape. This new paper they wanted to import into fast food companies. The paper is unique in that it is fully compostable and nongreasy. He asked me to create a new design for this type of product. We made an appointment with him in the office, where we met with five people - me and company representatives.
We consulted design parameters together. The design had to capture the advantages of paper and highlight the origin of the product. Due to the production price, it only had to be black and white. And for printing reasons, it only had to contain 100 percent stroke coverage. We clarified and agreed on the next meeting, when I presented them my first sketches.
I started exploring the websites. I was looking for inspiration in designs of packaging, but all things failed because of the colors. For a long time, I thought about how to highlight the benefits of paper with design. I solved this problem by designing vectors. I created five sketches, which I presented to the company's members at the next meeting. They consulted with each other and chose the direction in which we will continue.
The members of company liked the vektors and they wanted to use them in the design. I decided to create a cover background from vectors. I also started to focus on the main design. I didn't have much to choose from, so I opened an Adobe sketch and started drawing vector images.
The design should be playful and clean. For a long time I struggled with the vectors of individual dishes and their placement. In the end, I was inspired by the game of Five in Line - so the design was finally created. I presented the design to the company and the members were excited. Together we fine-tuned all the details and released the design to the printer.
Currently/nowadays this design is used in a large number of fast food restaurants in the Czech Republic. In the future, we plan to expand the supply to other countries like Slovakia, Poland or Germany, and I'll have to adjust/edit the design for these countries.